HBX-1992: Reorganize the reverse engineering strategy hierarchy - Rename 'org.hibernate.tool.internal.reveng.TableFilter' to 'org.hibernate.tool.internal.reveng.strategy.TableFilter'
