+++ orm/src/main/java/org/hibernate/tool/internal/reveng/strategy/OverrideBinder.java
@@ -15,7 +15,6 @@ import org.hibernate.tool.api.reveng.SchemaSelection;
 import org.hibernate.tool.api.reveng.TableIdentifier;
 import org.hibernate.tool.internal.reveng.DefaultAssociationInfo;
 import org.hibernate.tool.internal.reveng.SQLTypeMapping;
-import org.hibernate.tool.internal.reveng.TableFilter;
 import org.hibernate.tool.internal.reveng.strategy.MetaAttributeHelper.SimpleMetaAttribute;
 import org.hibernate.tool.internal.util.JdbcToHibernateTypeHelper;
 import org.w3c.dom.Document;
